# src/main/java/com/sprint/mission/discodeit/dto/data/MessageDTO.java
package com.sprint.mission.discodeit.dto.data;

import java.time.Instant;
import java.util.List;
import java.util.UUID;

public record MessageDTO(
    UUID id,
    Instant createdAt,
    Instant updatedAt,
    String content,
    UUID channelId,
    UserDTO author,
    List<BinaryContentDTO> attachments
) {

}